requirements.txt
pyyaml>=6.0
jinja2>=3.1
markdown>=3.5
rich>=13.0
click>=8.0